- Created new base class for duplicate detectors - Add factory class for duplicate manager to use to determine what detector to instantiate - Updated TE1 error message to align with expectations
