+++ tdrs-backend/tdpservice/data_files/test/test_api.py
@@ -138,8 +138,11 @@ class DataFileAPITestBase:
         assert sheet.cell(row=2, column=1).value == 202403
         assert sheet.cell(row=2, column=2).value == "*****5891"
         assert sheet.cell(row=2, column=3).value == 7
-        assert sheet.cell(row=2, column=4).value == ("Duplicate record detected with record type TE1 at line 7. "
-                                                     "Record is a duplicate of the record at line number 6.")
+        assert sheet.cell(row=2, column=4).value == ("Duplicate Social Security Number within a month. "
+                                                     "Check that individual Social Security Numbers within a "
+                                                     "single exit month are not included more than once. "
+                                                     "Social Security Number is a duplicate of the record at "
+                                                     "line number 6.")
 
         assert sheet.cell(row=3, column=1).value == 202301
         assert sheet.cell(row=3, column=2).value == "*****5507"
